FABMODO-20 now using int instead of unsigned int in the _polymesh structure (#1945)
* FABMODO-20 now using int instead of unsigned int in the _polymesh structure

* FABMODO-20 fix warnings 'comparison between signed and unsigned integer expressions'

* FABMODO-20 fix warnings 'comparison between signed and unsigned integer expressions' (cont)

* FABMODO-20 fix warnings 'comparison between signed and unsigned integer expressions' (cont)

+++ FabricDFGOperators.h
@@ -341,9 +341,9 @@ struct _polymesh
         polyNodeColors;
   */
 
-  unsigned int            numVertices;
-  unsigned int            numPolygons;
-  unsigned int            numSamples;
+  int                     numVertices;
+  int                     numPolygons;
+  int                     numSamples;
   std::vector <float>     vertPositions;
   std::vector <float>     vertNormals;
   std::vector <float>     vertUVWs;
@@ -414,11 +414,11 @@ struct _polymesh
     return (   numVertices >= 0
             && numPolygons >= 0
             && numSamples  >= 0
-            && vertPositions  .size() == 3 * numVertices
-            && vertNormals    .size() == 3 * numVertices
-            && polyNumVertices.size() ==     numPolygons
-            && polyVertices   .size() ==     numSamples
-            && polyNodeNormals.size() == 3 * numSamples
+            && (int)vertPositions  .size() == 3 * numVertices
+            && (int)vertNormals    .size() == 3 * numVertices
+            && (int)polyNumVertices.size() ==     numPolygons
+            && (int)polyVertices   .size() ==     numSamples
+            && (int)polyNodeNormals.size() == 3 * numSamples
            );
   }
 
@@ -431,13 +431,13 @@ struct _polymesh
   // returns true if this mesh has UVWs.
   bool hasUVWs(void) const
   {
-    return (vertUVWs.size() == 3 * numVertices && polyNodeUVWs.size() == 3 * numSamples);
+    return ((int)vertUVWs.size() == 3 * numVertices && (int)polyNodeUVWs.size() == 3 * numSamples);
   }
 
   // returns true if this mesh has Colors.
   bool hasColors(void) const
   {
-    return (vertColors.size() == 4 * numVertices && polyNodeColors.size() == 4 * numSamples);
+    return ((int)vertColors.size() == 4 * numVertices && (int)polyNodeColors.size() == 4 * numSamples);
   }
 
   // calculate bounding box (i.e. set member bbox).
@@ -454,7 +454,7 @@ struct _polymesh
       bbox[3] = pv[0];
       bbox[4] = pv[1];
       bbox[5] = pv[2];
-      for (unsigned int i=0;i<numVertices;i++,pv+=3)
+      for (int i=0;i<numVertices;i++,pv+=3)
       {
         bbox[0] = std::min(bbox[0], pv[0]);
         bbox[1] = std::min(bbox[1], pv[1]);
@@ -511,7 +511,7 @@ struct _polymesh
       // fill.
       uint32_t *pvi = polyVertices.data();
       float    *pnn = polyNodeNormals.data();
-      for (unsigned int i=0;i<numSamples;i++,pvi++,pnn+=3)
+      for (int i=0;i<numSamples;i++,pvi++,pnn+=3)
       {
         float *vn = vertNormals.data() + (*pvi) * 3;
         vn[0] += pnn[0];
@@ -521,7 +521,7 @@ struct _polymesh
 
       // normalize vertex normals.
       float *vn = vertNormals.data();
-      for (unsigned int i=0;i<numVertices;i++,vn+=3)
+      for (int i=0;i<numVertices;i++,vn+=3)
       {
         float f = vn[0] * vn[0] + vn[1] * vn[1] + vn[2] * vn[2];
         if (f > 1.0e-012f)
@@ -560,7 +560,7 @@ struct _polymesh
 
         // fill.
         uint32_t *pvi = polyVertices.data();
-        for (unsigned int i=0;i<numSamples;i++,pvi++)
+        for (int i=0;i<numSamples;i++,pvi++)
           tmpVertNumPolyNeigh.data()[*pvi]++;
       }
     }
@@ -582,7 +582,7 @@ struct _polymesh
       // fill.
       uint32_t *pvi = polyVertices.data();
       float    *pnu = polyNodeUVWs.data();
-      for (unsigned int i=0;i<numSamples;i++,pvi++,pnu+=3)
+      for (int i=0;i<numSamples;i++,pvi++,pnu+=3)
       {
         float *vu = vertUVWs.data() + (*pvi) * 3;
         vu[0] += pnu[0];
@@ -593,7 +593,7 @@ struct _polymesh
       // normalize.
       uint32_t *tn = tmpVertNumPolyNeigh.data();
       float    *vu = vertUVWs.data();
-      for (unsigned int i=0;i<numVertices;i++,tn++,vu+=3)
+      for (int i=0;i<numVertices;i++,tn++,vu+=3)
       {
         if (*tn > 1)
         {
@@ -622,7 +622,7 @@ struct _polymesh
       // fill.
       uint32_t *pvi = polyVertices.data();
       float    *pnc = polyNodeColors.data();
-      for (unsigned int i=0;i<numSamples;i++,pvi++,pnc+=4)
+      for (int i=0;i<numSamples;i++,pvi++,pnc+=4)
       {
         float *vc = vertColors.data() + (*pvi) * 4;
         vc[0] += pnc[0];
@@ -634,7 +634,7 @@ struct _polymesh
       // normalize.
       uint32_t *tn = tmpVertNumPolyNeigh.data();
       float    *vc = vertColors.data();
-      for (unsigned int i=0;i<numVertices;i++,tn++,vc+=4)
+      for (int i=0;i<numVertices;i++,tn++,vc+=4)
       {
         if (*tn > 1)
         {
@@ -687,10 +687,10 @@ struct _polymesh
     numSamples  = in_polyVertices_size;
 
     // check.
-    if (    numVertices * 3 != in_vertPositions_size
-        || (numSamples  * 3 != in_nodeNormals_size && in_nodeNormals_size > 0)
-        || (numSamples  * 3 != in_nodeUVWs_size    && in_nodeUVWs_size    > 0)
-        || (numSamples  * 4 != in_nodeColors_size  && in_nodeColors_size  > 0))
+    if (    (unsigned int)numVertices * 3 != in_vertPositions_size
+        || ((unsigned int)numSamples  * 3 != in_nodeNormals_size && in_nodeNormals_size > 0)
+        || ((unsigned int)numSamples  * 3 != in_nodeUVWs_size    && in_nodeUVWs_size    > 0)
+        || ((unsigned int)numSamples  * 4 != in_nodeColors_size  && in_nodeColors_size  > 0))
     {
       clear();
       return -2;
@@ -726,7 +726,7 @@ struct _polymesh
       // vertex positions.
       double *src = (double *)in_vertPositions;
       float  *dst = vertPositions.data();
-      for (unsigned int i=0;i<numVertices;i++,src+=3,dst+=3)
+      for (int i=0;i<numVertices;i++,src+=3,dst+=3)
       {
         dst[0] = (float)src[0];
         dst[1] = (float)src[1];
@@ -754,7 +754,7 @@ struct _polymesh
         unsigned int *pn  = polyNumVertices.data();
         unsigned int *pi  = polyVertices   .data();
         float        *pnn = polyNodeNormals.data();
-        for (unsigned int i=0;i<numPolygons;i++,pn++)
+        for (int i=0;i<numPolygons;i++,pn++)
         {
           if (*pn <= 2)
           {
@@ -826,7 +826,7 @@ struct _polymesh
       // fill.
       uint32_t *pvi = polyVertices.data();
       float    *pnn = polyNodeNormals.data();
-      for (unsigned int i=0;i<numSamples;i++,pvi++,pnn+=3)
+      for (int i=0;i<numSamples;i++,pvi++,pnn+=3)
       {
         float *vn = vertNormals.data() + (*pvi) * 3;
         vn[0] += pnn[0];
@@ -836,7 +836,7 @@ struct _polymesh
 
       // normalize vertex normals.
       float *vn = vertNormals.data();
-      for (unsigned int i=0;i<numVertices;i++,vn+=3)
+      for (int i=0;i<numVertices;i++,vn+=3)
       {
         float f = vn[0] * vn[0] + vn[1] * vn[1] + vn[2] * vn[2];
         if (f > 1.0e-012f)
@@ -873,7 +873,7 @@ struct _polymesh
       // fill.
       uint32_t *pvi = polyVertices.data();
       float    *pnu = polyNodeUVWs.data();
-      for (unsigned int i=0;i<numSamples;i++,pvi++,pnu+=3)
+      for (int i=0;i<numSamples;i++,pvi++,pnu+=3)
       {
         float *vu = vertUVWs.data() + (*pvi) * 3;
         vu[0] += pnu[0];
@@ -885,7 +885,7 @@ struct _polymesh
       // average the vertex values.
       float     *vu = vertUVWs.data();
       short int *vt = tmp.data();
-      for (unsigned int i=0;i<numVertices;i++,vu+=3,vt++)
+      for (int i=0;i<numVertices;i++,vu+=3,vt++)
       {
         if (*vt > 0)
         {
@@ -915,7 +915,7 @@ struct _polymesh
       // fill.
       uint32_t *pvi = polyVertices.data();
       float    *pnc = polyNodeColors.data();
-      for (unsigned int i=0;i<numSamples;i++,pvi++,pnc+=4)
+      for (int i=0;i<numSamples;i++,pvi++,pnc+=4)
       {
         float *vc = vertColors.data() + (*pvi) * 4;
         vc[0] += pnc[0];
@@ -928,7 +928,7 @@ struct _polymesh
       // average the vertex values.
       float     *vc = vertColors.data();
       short int *vt = tmp.data();
-      for (unsigned int i=0;i<numVertices;i++,vc+=4,vt++)
+      for (int i=0;i<numVertices;i++,vc+=4,vt++)
       {
         if (*vt > 0)
         {
@@ -1024,7 +1024,7 @@ struct _polymesh
 
     // fix vertex indices.
     uint32_t *pi = polyVertices.data() + numSamples;
-    for (unsigned int i=0;i<inMesh.numSamples;i++,pi++)
+    for (int i=0;i<inMesh.numSamples;i++,pi++)
       *pi += numVertices;
 
     // fix amounts.
